v.3

+++ ram.h
@@ -1,21 +1,20 @@
 #ifndef RAM_H
 #define RAM_H
 
+#include <stddef.h>
+
 typedef struct RAM {
-  int *memory;
+  int* memory;
   size_t size;
 } RAM;
 
-// create ram
 RAM* create_ram(size_t size);
 RAM* create_empty_ram(size_t size);
 RAM* create_random_ram(size_t size);
 
-// get, set and print_ram
-int get_ram(RAM *ram, int memory_address);
-void set_ram(RAM *ram, size_t memory_address, int new_memory_value);
-//void print_ram(RAM *ram);
+int get_ram(RAM* ram, size_t memory_address);
+void set_ram(RAM* ram, size_t memory_address, int new_memory_value);
 
-void destroy_ram(RAM *ram);
+void destroy_ram(RAM* ram);
 
-#endif // !RAM_H
+#endif
